Extracted fields packages: dropdown, radio, text
