+++ ui-core/src/main/kotlin/com/processout/sdk/ui/core/component/field/text/POLabeledTextField.kt
@@ -1,4 +1,4 @@
-package com.processout.sdk.ui.core.component.field
+package com.processout.sdk.ui.core.component.field.text
 
 import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Column
@@ -14,6 +14,9 @@ import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.processout.sdk.ui.core.annotation.ProcessOutInternalApi
+import com.processout.sdk.ui.core.component.field.LabeledFieldLayout
+import com.processout.sdk.ui.core.component.field.POField
+import com.processout.sdk.ui.core.component.field.POFieldLabels
 
 /** @suppress */
 @ProcessOutInternalApi
